# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A control plane for Chorus Worker that manages S3 replication jobs",
        "title": "Chorus Controller API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/buckets": {
            "get": {
                "description": "Returns a list of buckets that can be used for replication jobs between specified storages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buckets"
                ],
                "summary": "List buckets available for replication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User identifier",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Source storage name",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Destination storage name",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to show already replicated buckets",
                        "name": "show_replicated",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/storages": {
            "get": {
                "description": "Returns a list of all configured storage backends",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storages"
                ],
                "summary": "List all storages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Creates a storage with simplified parameters. All other settings use default values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storages"
                ],
                "summary": "Create a storage configuration",
                "parameters": [
                    {
                        "description": "Storage creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateStorageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Storage created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/storages/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storages"
                ],
                "summary": "Get storage by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Storage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Updates an existing storage with new parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storages"
                ],
                "summary": "Update a storage configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Storage update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateStorageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Storage updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Deletes a storage and all its credentials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storages"
                ],
                "summary": "Delete a storage configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Storage deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.CreateStorageRequest": {
            "type": "object",
            "required": [
                "access_key",
                "address",
                "name",
                "provider",
                "secret_key",
                "user"
            ],
            "properties": {
                "access_key": {
                    "type": "string",
                    "example": "AKIA123"
                },
                "address": {
                    "type": "string",
                    "example": "http://localhost:9000"
                },
                "name": {
                    "type": "string",
                    "example": "my-storage"
                },
                "provider": {
                    "type": "string",
                    "example": "minio"
                },
                "secret_key": {
                    "type": "string",
                    "example": "SECRET123"
                },
                "user": {
                    "type": "string",
                    "example": "myuser"
                }
            }
        },
...
        "domain.Storage": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "default_region": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "health_check_interval_ms": {
                    "type": "integer"
                },
                "http_timeout_ms": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "is_secure": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rate_limit_enabled": {
                    "type": "boolean"
                },
                "rate_limit_rpm": {
                    "type": "integer"
                },
                "secret_access_key": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "description": "Type \"Token\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}